backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.10
mlflow==3.3.2
pydantic==2.11
pandas==2.3.2
scikit-learn==1.5.2
numpy==2.3.2
xgboost==3.0.4
PyJWT==2.10.1
bcrypt==4.3.0
redis==6.0.0
minio==7.2.15